+++ src/Twilio/Rest/Preview/Sync/Service/DocumentOptions.cs
@@ -65,10 +65,6 @@ namespace Twilio.Rest.Preview.Sync.Service
         /// The sid
         /// </summary>
         public string PathSid { get; }
-        /// <summary>
-        /// The If-Match HTTP request header
-        /// </summary>
-        public string IfMatch { get; set; }
 
         /// <summary>
         /// Construct a new DeleteDocumentOptions
@@ -87,20 +83,6 @@ namespace Twilio.Rest.Preview.Sync.Service
         public List<KeyValuePair<string, string>> GetParams()
         {
             var p = new List<KeyValuePair<string, string>>();
-            return p;
-        }
-
-        /// <summary>
-        /// Generate the necessary header parameters
-        /// </summary>
-        public List<KeyValuePair<string, string>> GetHeaderParams()
-        {
-            var p = new List<KeyValuePair<string, string>>();
-            if (IfMatch != null)
-            {
-                p.Add(new KeyValuePair<string, string>("If-Match", IfMatch));
-            }
-
             return p;
         }
     }