+++ CHANGES.md
@@ -1,6 +1,40 @@
 twilio-csharp Changelog
 =======================
 
+[2021-02-12] Version 5.54.0
+---------------------------
+**Library - Chore**
+- [PR #554](https://github.com/twilio/twilio-csharp/pull/554): update documentation in upgrade.md. Thanks to [@JenniferMah](https://github.com/JenniferMah)!
+
+**Api**
+- Revert change that conference participant create `from` and `to` param to be endpoint type for supporting client identifier and sip address
+- Update the conference participant create `from` and `to` param to be endpoint type for supporting client identifier and sip address
+
+**Events**
+- Documentation should state that no fields are PII
+
+**Flex**
+- Adding `notifications` and `markdown` to Flex Configuration
+
+**Messaging**
+- Add A2P use cases API
+- Add Brand Registrations API
+- Add Campaigns API
+
+**Serverless**
+- Add runtime field to Build response and as an optional parameter to the Build create endpoint.
+- Add @twilio/runtime-handler dependency to Build response example.
+
+**Sync**
+- Remove If-Match header for Document **(breaking change)**
+
+**Trusthub**
+- Make all resources public
+
+**Twiml**
+- Add `refer_url` and `refer_method` to `Dial`.
+
+
 [2021-01-27] Version 5.53.1
 ---------------------------
 **Studio**